hit module improvement to forgetting_factor computing with pow, and other chanes to level module
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3257 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ modules/extra/webp/samson/modules/webp/CategoryHits.h
@@ -41,7 +41,7 @@ namespace webp{
 		  // Apply a forgetting factor to "current" counter
 		  double time_span = 300;
 		  double factor = ( time_span - 1 ) / time_span;
-		  current.value = current.value * pow( factor  , seconds );		  
+		  current.value = current.value * pow( factor  , (int) seconds );
 	   }
